[AN] chore: ProfileContent 파일 분리 #721
