+++ android/app/src/main/java/com/onair/hearit/presentation/setting/component/ProfileTopBar.kt
@@ -1,0 +1,49 @@
+package com.onair.hearit.presentation.setting.component
+
+import androidx.compose.material3.CenterAlignedTopAppBar
+import androidx.compose.material3.ExperimentalMaterial3Api
+import androidx.compose.material3.Icon
+import androidx.compose.material3.IconButton
+import androidx.compose.material3.Text
+import androidx.compose.material3.TopAppBarDefaults
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.res.painterResource
+import androidx.compose.ui.res.stringResource
+import androidx.compose.ui.tooling.preview.Preview
+import com.onair.hearit.R
+import com.onair.hearit.presentation.theme.Gray4
+import com.onair.hearit.presentation.theme.HearitBlack
+import com.onair.hearit.presentation.theme.HearitTypoGraphy
+
+@OptIn(ExperimentalMaterial3Api::class)
+@Composable
+fun ProfileTopBar(onBackClick: () -> Unit) {
+    CenterAlignedTopAppBar(
+        title = {
+            Text(
+                text = stringResource(R.string.setting_profile),
+                style = HearitTypoGraphy.titleLarge,
+            )
+        },
+        navigationIcon = {
+            IconButton(onClick = onBackClick) {
+                Icon(
+                    painter = painterResource(id = R.drawable.ic_back),
+                    contentDescription = "뒤로가기",
+                )
+            }
+        },
+        colors =
+            TopAppBarDefaults.centerAlignedTopAppBarColors(
+                containerColor = HearitBlack,
+                navigationIconContentColor = Gray4,
+                titleContentColor = Gray4,
+            ),
+    )
+}
+
+@Preview
+@Composable
+fun ProfileTopBarPreview() {
+    ProfileTopBar(onBackClick = {})
+}
